requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.14.0
jwcrypto==1.3.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
